% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cor_est.R
\name{DGCFit}
\alias{DGCFit}
\title{Fit discrete multivariate data using Gaussian copula}
\usage{
DGCFit(mapped_data, Fhats, nearPDproj = T, ncores = 2, ...)
}
\arguments{
\item{mapped_data}{A discrete dataset whose discrete values are all numeric or have been mapped to numbers.}

\item{nearPDproj}{Logical. If TRUE, the estimated correlation matrix using Kendall's tau will be projected to the nearest positive definite matrix.}

\item{ncores}{Number of cores to use when parallelized computing of the correlation matrix.}

\item{Fhat}{A list of estimated marginal cdfs of the dataset. Each item in the list is a \code{stepfun} object and the support of the cdfs (\code{knots(Fhats[[j]])}) should be the superset of the unique discrete values of the corresponding feature (\code{unique(mapped_data[,j])}). If not specified, \code{ecdf} will be used to estimate the marginal cdfs.}
}
\value{
An object of class \code{DGCFit} with components
\item{mapped_data}{The input \code{mapped_data}.}
\item{tau}{The Kendall's tau matrix of the input discrete numerical dataset.}
\item{corr}{The estimated correlation matrix using the Kendall's tau matrix.}
\item{cdfs}{The list of estimated marginal cdfs.}
}
\description{
Fit discrete multivariate data using Gaussian copula
}
