% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dgcpca.R
\name{DGCPCA}
\alias{DGCPCA}
\alias{DGCPCA.DGCFit}
\alias{DGCPCA.default}
\title{Perform PCA for discrete dataset using Gaussian copula}
\usage{
DGCPCA(x, npc, ...)

\method{DGCPCA}{DGCFit}(x, npc = 5, ...)

\method{DGCPCA}{default}(x, npc = 5, maps = TRUE, ...)
}
\arguments{
\item{x}{A discrete dataset. Assumed to be all numerical-discrete-valued if \code{maps = FALSE}.}

\item{npc}{Number of first PCs needed.}

\item{...}{Parameters for \code{\link{DGCFit}}.}

\item{maps}{Logical or a list of vectors of length \code{ncol(x)}. If FALSE, \code{x} must be a mapped dataset (i.e. all discrete values are numerical). If TRUE, all values will be mapped to discrete values in [0,1] using \code{sort(unique(x[,j]))}. If a list of vectors of length \code{ncol(x)}, values will be mapped to discrete values in [0,1] accoording to the provided list. See \code{\link{discrete_mapping}}.}
}
\value{
An object of class \code{DGCPCA} with components:
\describe{
\item{fit}{An object of \code{DGCFit}. See \code{\link{DGCFit}}.}
\item{npc}{Number of first PCs needed.}
\item{lambdas}{Eigenvalues for the first \code{npc} PCs.}
\item{rotmat}{Rotation matrix.}
\item{scores}{The first \code{npc} surrogate PC scores.}
\item{latent_norm}{The simulated unobserved multivariate normal random vectors used to generate the surrogate PC scores.}
}
}
\description{
Perform PCA for discrete dataset using Gaussian copula
}
\section{Methods (by class)}{
\itemize{
\item \code{DGCFit}: S3 method for \code{DGCFit} object

\item \code{default}: S3 method for default
}}

