% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copula_fit.R
\name{discrete_mapping}
\alias{discrete_mapping}
\title{Map every column of the original discrete data to new data with discrete values in [0,1]}
\usage{
discrete_mapping(data, maps)
}
\arguments{
\item{data}{The original discrete data}

\item{maps}{A list of the same length of ncol(data), each item is a vector containing the original image of the map. The vector should represent the ascending order of the discrete values in the original data, as the discrete values will be mapped to numbers in [0,1] in a ascending order.}
}
\value{
A new dataset. All values are in [0,1].
}
\description{
Map every column of the original discrete data to new data with discrete values in [0,1]
}
