% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copula_fit.R
\name{gdtrans}
\alias{gdtrans}
\title{Generalized distributional transformation of discrete data}
\usage{
gdtrans(mapped_data, Fhats)
}
\arguments{
\item{mapped_data}{A discrete dataset whose discrete values are all numeric or have been mapped to numbers.}

\item{Fhat}{A list of estimated marginal cdfs of the dataset. Each item in the list is a \code{stepfun} object and the support of the cdfs (\code{knots(Fhats[[j]])}) should be the superset of the unique discrete values of the corresponding feature (\code{unique(mapped_data[,j])}). If not specified, \code{ecdf} will be used to estimate the marginal cdfs.}
}
\description{
Generalized distributional transformation of discrete data
}
