% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_recovery.R
\name{get_recovery_accuray}
\alias{get_recovery_accuray}
\title{Get the recovery accuracies}
\usage{
get_recovery_accuray(dgcpca)
}
\arguments{
\item{dgcpca}{An object of class \code{DGCPCA}.}
}
\value{
A vector of recovery accuracies.
}
\description{
Get the recovery accuracies using 0 to \code{dgcpca$npc} PCs.
}
